Make Pod::Simple::XHTML compatible with Pod::Simple::HTMLBatch
The first change is to add the empty method html_header_after_title().
HTMLBatch requires this method.  For now, I've left it empty and this seems to
work fine.

The second change is to modify start_Document() to not put the value of
html_css() into a <link> tag if its already in a <link> tag.  When HTMLBatch
calls html_css() it passes in a <link> tag rather than just a plain url.