Allow directives that Pod::Simple autocloses to be recognized in subclasses.
=end/=back directives now pass an attribute hash ref, similar to the
hash ref passed in all _handle_element_start() calls, to
_handle_element_end(). This allows subclasses to see if there exists
a key 'fake-closer' which would signify that Pod::Simple generated the
closing directive and it is not really present in the POD source.

Pod::Simple::Methody has been updated to pass the hash ref through too.

Included is the updated code, documentation, and tests.