Don't end XHTML IDs with punctuation
An ID ending with a punctuation character means that a URL with a fragment
ID linking to that section ends with a punctuation character. Such URLs
are awkward to extract from plain text, for example when used in a plain
text e-mail and being converted to a link by a mailing list archive or
e-mail client: a punctuation character at the end of a URL may be
interpreted as punctuation in the e-mail message following the URL.

Avoid the problem by ensuring that IDs never end with punctuation
characters. (Underscore is allowed, since it isn't used for punctuation in
human text.)