whine when L<> starts or ends with whitespace
For example:

  L< Foo>
  L<Foo >

It leads to confusion.  For example, L<Foo> is a link to the module
Foo, but L< Foo> is a link to the section " Foo" in this document.

This check was in Pod::Checker.  In fact, Pod::Parser was resolving
things the *other* way!  It trimmed whitespace first!