encod04.t: Restructure 2 tests to be TODOs
These tests pass when the tests actually should fail.  Instead make them
TODO.

I came up with a somewhat less contrived example for the first one,
where it could legally be either UTF-8 or CP1252.  This gives me hope
that real-world examples will be quite unlikely to be guessed wrong.
The first byte must be between C2 and DF, otherwise it would be a 3 byte
sequence in UTF-8, and even harder to find a likely CP1252 equivalent
sequence.  That means that the first byte is an uppercase accented
character, the multiplication sign, or the German sharp s 'ß'.  The
second byte is in the range 80 to 9F.  Most of these are various
punctuation characters or symbols such as a dagger.  These are mostly
unlikely to immediately follow an uppercase letter, or the sharp s.  One
that could is a right single quote used as an apostrophe in English.
But there are no accents in English except in borrowed words.  Since it
must be a capital, it's likely the whole word is in caps, like in a
heading.  I came up with what looks like "JOSÉ'S" in CP1252, which looks
like legal UTF-8 as well.