encod04.t: Fix-up two tests
One test has been failing because it was testing that illegal UTF-8 was
considered to be UTF-8.  This commit fixes that.

The other test is made a TODO.  It is passed genuninely ambiguous text
that could either be CP1252 or UTF-8.  This commit makes the text passed
actually more plausible than previously.  The fact that it was hard to
get a plausible example gives me hope that real-world examples will be
quite unlikely to be guessed wrong.  The first byte must be between C2
and DF, otherwise it would be a 3 byte sequence in UTF-8, and even
harder to find a likely CP1252 equivalent sequence.  That means that the
first byte is one of 1) an uppercase accented character, 2) the
multiplication sign, or 3) the German sharp s 'ß'.  The second byte is
in the range 80 to 9F.  Most of these in CP1252 are various punctuation
characters or symbols such as a dagger.  These are mostly unlikely to
immediately follow an uppercase letter, multiplcation sign, or the sharp
s.  One that could is a right single quote used as an apostrophe in
English.  But there are no accents in English except in borrowed words.
Since it must be a capital, it's likely the whole word is in caps, like
in a heading.  I came up with what looks like "JOSÉ'S" in CP1252, which
looks like legal UTF-8 as well.