Blackbox.pm: Improve encoding detection, extend to EBCDIC
The spec says that a pod without an =encoding line is to be considered
to be ISO-8859-1 (Latin1) or UTF-8, based on its contents.  As long as
the pod contains only characters that are the same in both encodings, we
don't have to choose between the two encodings.  But the choice does
have to be made before processing the first line that contains a byte
that has a different meaning depending on the encoding.

The heuristic this commit replaces worked pretty well on ASCII
platforms, but not perfectly, and not at all on EBCDIC.  This commit
adds some rigor and extends it to work on EBCDIC.  The only thing I can
think of to improve the heuristic in this commit would be to add an
extra pass through the pod to look for more evidence.  That would be
overkill.

The crux of the problem is about bytes whose meaning varies depending on
whether the encoding is UTF-8 or not.  These are called UTF-8 variants.
On ASCII platforms, the set of UTF-8 variants is identical to the set of
non-ASCII characters, so the two concepts are often conflated.  But this
is not the case on EBCDIC platforms, and conflating them leads to wrong
code.

The previous heuristic looked at just the first sequence of one or two
variant bytes on the first input line containing variants, and if those
looked like valid UTF-8, called the whole file UTF-8.  It threw away any
extra information.  This actually works quite well on ASCII platforms,
because anything that is legal UTF-8 is quite unlikely to appear in real
8859-1 text just because of what characters are assigned where.  But
cp1252 text can be mistaken, rather disastrously, for UTF-8.  It would
be far less disastrous to make the opposite choice.

This commit looks at all the variant sequences on the line, and at all
the bytes in each sequence.  It doesn't throw away any information.  If
the line contains just a single sequence of two bytes, we are really no
better off than before this commit, but if it contains longer and/or
more sequences, our accuracy goes way up.

This code does add extra rigor to the previous, so some two-byte
sequences that were previously accepted as looking like UTF-8 are now
rejected because they aren't actually UTF-8.

This commit causes one TODO test to pass.