Fix encoding guessing to work on EBCDIC platforms
When no =encoding line is present, the encoding is checked to see if it
is UTF-8, and if not, currently ISO 8859-1 is chosen instead.  This
wasn't working well on EBCDIC platforms prior to this commit.

It is planned to change things so that CP 1252 is chosen instead of
8859-1, and this code will have to be revised to handle that, but in
case that doesn't work out, this commit can be fallen back to.