Generalize XHTML name detection for non-ASCII platforms
This commit takes two identical regular expression patterns and makes
them into a single qr//.  And it rewrites the revised one so it is
platform-independent on sufficiently modern Perls.

I think the pattern is wrong to exclude the digit '9', but I don't have
time now to develop the expertise to delve into it, so am leaving it
as-is.  I compiled the two versions under -Dr (one using hard-coded
characters, and the other using [:posix:] classes) to verify that the
new one generates the exact same code points as the original on ASCII
platforms