HTML.pm: Fix bad regex pattern
The :^ascii: should be part of a bracketed character class.  I missed
this in code review.  There is code in regcomp.c to warn on something
like this, but it didn't get triggered, I'll look into that.  And I
didn't add a test for this.  It's not critical if such characters don't
get dropped.