Fix cp1252 for early Perls
Perls prior to v5.8 did not work well with UTF-8.  This presents
problems for the CP1252 character encoding, as some of its code points,
when converted to Unicode, require UTF-8 to represent.  This patch
instead uses ASCII approximations for them.  Prior to this patch, they
were left alone, which would show up as C1 control characters.

On Perls that do work with UTF-8, the code points are still properly
converted to their Unicode equivalents.