Assume .pod is always Pod.
Even if contains_pod() returns false. This is to make it more consistent
with survey(), which makes that assumption. Fixes a test failure when
Forward::Routes is installed, because 'Forward/Guides/Routes/BlockStyle.pod'
does not actually contain Pod.

It might be that we want to tighten this down, but since survey() predates
find(), it seems better to be consistent with the older code, at least for
now.