Add case-insensitive and -preservation smarts to Search.
The survey() method was faling to ignore dupes on case-insensive file systems
where the case of file names differed. So add a special case-insensitive check
for duplicates on case-sensitive file systems. The heurisitic for determining
if a file system is case-insensitive is currently a bit of a hack, but should
work in most cases.

Also, when the find() nethod searches a "./pod" subdirectory, preserve the
case of that directory in the results. This is done by, unfortunately, reading
every file in the directory to find any that match `lc 'pod'`, which might be
slow. Fortunately it doesn't usually search very many directories. This fixes
test failures on Windows.