Better handle searching for Pod::*.
Fix the find() method in Pod::Simple::Search to find the proper file on
case-insensitive file systems when searching for modules starting with "Pod".
For example, when searching for 'Pod::Perldoc' it now returns a file ending in
'Pod/Perldoc.pm' instead of 'Pod/perldoc.pod', as the latter is actually the
documention for the 'perldoc' program. This fixes a test failure on Windows.