Allow search current directory in find() and survey().
Prior to v3.30, it was excluded by survey() but not by find. I can find no
documentation as to why this might be so, but it does seem to have avoided the
appearance of duplicate directories, especially when testing and PERL5LIB
might duplicate the directory. So instead filter for dupes.