Extend CP1252/UTF-8 script detection for early perls
Before Perl 5.6, there was no UTF-8, so must be CP1252.

On perls that don't have utf8::decode(), this commit implements that
function (as far as needed for this module) in pure perl.

It changes from looping through the line looking for segments that are
non-ASCII and then examining them in more detail, to just look at the
line as a whole if we have the decode function.  And if not, we then
look at individual bytes to see if it is legal UTF-8.

This removes a TODO for a test on early perls