Add more heuristics for guessing between CP1252/UTF-8
This adds a bunch more criteria to improve our guess of what script a
pod is in if it contains more than ASCII and no =encoding line.

One criteria added checks for any of the bytes not defined by CP1252.
If these are found, and it is syntactically legal UTF-8, guess UTF-8.

The other tests basically exclude UTF-8 if one of the code points
found would unlikely be found in pods.  This includes C1 controls,
unassigned code points, surrogates, deprecated code points,
International Phonetic Alphabet characters, and Latin characters not in
the first release of Unicode.  This last is because Unicode included in
its first release virtually every Latin character in modern usage.

These criteria are all I could think of (except for the ones in the
following commit) that are reasonably testable.  Some of them are
testable even in Perl 5.6.  And those are written so that they do
compile on 5.6.

A perl that doesn't have the ability to use one of the given criteria
simply skips using it, giving graceful degradation.

This removes a TODO from one of the tests.