Add mixed script detection for CP1252/UTF-8 choice
When trying to guess whether a pod without an =encoding line is CP1252
or UTF-8, this commit looks for mixed scripts.  CP1252 is all Latin1
(plus punctuation, etc), whereas UTF-8 can be anything.

But pods will tend to all be written in the same language and script, so
if our ambiguous input text shows as multiple scripts if it were encoded
in UTF-8, it's far more likely to be CP1252.

Perl 5.28 adds script runs to detect mixed scripts.  This commit uses
that feature if available.  When run on earlier perls, what can be done
is see if there are both Latin and non-Latin characters present.  This
isn't as good as the script run, but it catches most of the issues.