JustPod: Use correct white space in eg C<< foo >>
This commit changes to use the actual white space that was in the file
being parsed between the opening and closing brackets and the real
interior of pod formatting codes.  This is now saved for our use by
Blackbox.pm.

The white space that is after the interior and before the closing
brackets is now saved on a stack that is pushed as the beginning of the
construct is parsed, and popped as the trailing part is parsed.