JustPod01.t: Allow to work in the absence of Text::Diff
Prior to this commit Text::Diff was required to be installed to get this
test to even compile.  Now, it uses PERL_DIFF_TOOL, if set, to show the
differences.  Otherwise, Text::Diff is used if available, but the
absence of that module works, and the output is less than optimal for
any failures found.