PATCH: github issue 79: =cut event out of order
https://github.com/perl-pod/pod-simple/issues/79

When it sees a =cut, the code tries to finish up the current paragraph,
and then calls the cut handler.  However, the finishing up doesn't
actually do anything for 3 kinds of paragraphs, one of which is
verbatim, as is failing in this ticket.

The comments indicate that this is because these paragraph types need to
do look ahead.  But in the case of the =cut, it has terminated this
portion of the pod, and so there is nothing to look ahead to.

This commit solves the problem by examining these types of paragraphs
after the =cut, which has set in_pod to 0.