RTF.pm: Generate unicode characters again
This resolves github issue #92.

The problem was on all but ancient perls, a pattern was matching
everything but ascii characters, and it was supposed to stop at \xff.
I think it was me who introduced this bug at some point, and since we
didn't have a test file that included Unicode RTF, the bug went
unnoticed until reported.  Code existed in this module to handle the
case; it's just that this pattern caused all above-255 code points to be
deleted from the input before that code got executed.

The solution here is to not use that pattern at all, but to generate a
new pattern based on just the characters that we have escapes for.  This
also works on EBCDIC platforms of all vintages.

However, I notice that the code doesn't handle code points above 0xFFFF.
This will be addressed in the next few commits.