Don't return that content was seen unless this is pod
Previously a garbage file, as evidenced by the test case introduced, a
pgp key, when passed to Pod::Simple could indicate it was pod.  Now, it
really has to have a valid pod command in it for the content_seen method
to return true

This fixes a portion of Issue #95

There may be more needed due to other ways of installing new directives,
but rather than learning the code thoroughly now, I'll wait until a bug
presents itself.