implement images in POD
The patch allows pod formatters to recognize new image targets,
and include these images in the output. The new image target is
described as YAML, such as:

  =begin image

  src:image.png
  title: I<Fig.1>

  =end image

or simply

  =for image src:image.png

Both older parsers and the formatters that are not aware of the new
functionality will retain full backward compatibility.  There are added
extra provisions for them, to render POD instead of images:

  =for image-text src:image.png

  I<text fallback>

  =for image-cut

See more:

- lib/Pod/Simple/Images.pod - syntax extensions
- lib/Pod/Simple/HTML.pm    - image-aware HTML formatters
- lib/Pod/Simple/XHTML.pm
- lib/Pod/Simple/YAML.pm    - YAML 1.2 parser