clean up path handling in tests
The tests have a lot of cruft due to the history of this dist. They
attempt to check multiple locations based on if testing is being done in
core or not, and trying to account for different starting directories.
They change directory in various places which can be confusing.

This is now dual-life, and is packaged in core's `cpan/` directory. And
core's testing infrastructure has been improved over time. Most of the
extra tricks thus no longer required. Instead, we can find the files and
directories based on the path of the test script, and don't need to
bother trying to add lib paths or caring about PERL_CORE. We can also
avoid changing directory.