remove optional use of Test::Deep
One test script uses Test::Deep optionally. Test::Deep is a non-core
module so it is a poor choice for testing a dual-life module. It also
isn't being used for anything particularly useful. It can be trivially
be replaced with a simple string equality check.

One use of print as a method is also removed, as it requires IO::File
to be loaded. That would implicitly be done by Test::Deep, but we can
avoid the issue by just using the print function.