refactor link handling to encode in correct locations
In Pod::Simple::XHTML, the places where entity encoding for heading
links and targets happen are not really correct, and only work because
the id conversion strips many characters from them. It also lacks any
URL encoding for section links, which again only works because of the
stripping.

If a subclass tries to change how the id conversion happens, all of the
encoding will be wrong. Properly handling the encoding requires doing
the id conversion before any encoding, and needs different encoding for
targets vs links.

This refactors the code, moving when encoding happens. This removes the
need to strip tags and entities from the id conversion process, although
some additional stripping is needed to maintain compatibility. Section
links are also URL encoded. The output should be identical, but it
allows a subclass to modify the idify method and still produce proper
links.