test encoding of ids/links with an extension allowing more flexible ids
The encoding of ids and fragment links was fixed, but it would have no
impact on normal operations because of the aggressive filtering of ids.

Add a test which extends Pod::Simple::XHTML to do less filtering on the
generated ids. This type of id will only work properly with the new,
fixed, encoding.