XHTML: =over without =item should render as a blockquote
If an =over section doesn't include =item directives under it, it is
meant to be interpreted as an indented section or block quote. Producing
a <ul> element with something other than <li> inside it is invalid HTML.

Change the XHTML formatter to produce a <blockquote> element for this
type of =over section, matching the HTML formatter and the Pod spec's
recommendation.