rename select method to set_heading_select to avoid conflicts
Pod::Usage subclasses Pod::Simple and implements its own select method.
That method tries to implement its own behavior but if a select method
exists in the parent it will call that instead. This is for
compatibility with inheriting from Pod::Parser instead.

Avoid that compatibility issue by renaming the select method. It has
only been available in a stable release for a few days and isn't yet
documented, so it's pretty unlikely anyone is using it yet.